requirements.txt
Django
gunicorn
whitenoise
pyrebase4
